oled-tv/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "oled-tv"
version = "0.1.0"
description = "OLED Desktop Mirror Driver"
requires-python = ">=3.8"

dependencies = [
  "pyserial",
  "mss",
  "Pillow"
]

[project.scripts]
oled-tv = "oled_tv.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]